fixing delete ssues

+++ actions/delete-post.ts
@@ -6,7 +6,14 @@ import { redirect } from "next/navigation"
 
 
 export const DeletePost = async (postId : number) => {
-    const supabase = await createClient()
+    const supabase = await createClient() 
+
+      // Check if user is authorized
+
+    const { data: { user } } = await supabase.auth.getUser()
+
+    if (!user) throw new Error("Not authorized")
+        
     await supabase
                 .from("posts")
                 .delete()
